requirements.txt
requests
beautifulsoup4
pdfplumber
camelot-py[cv]
pandas
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib